chore: refactoring code into more classes etc.
