+++ src/main/scala/com/funny/utils/WebData.scala
@@ -1,0 +1,5 @@
+package com.funny.utils
+
+import scala.concurrent.Future
+
+case class WebData(webSiteData: Future[WebSiteData])
